requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
discord.py==2.6.4
distro==1.9.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
multidict==6.7.0
openai==2.5.0
propcache==0.4.1
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
